requirements.txt
pystray>=0.19
Pillow>=9.0
pynput>=1.7
